requirements.txt
# PACoS Brain Dependencies
uagents>=0.4.0
pydantic>=2.0.0
anthropic>=0.7.0
google-api-python-client>=2.100.0
python-dotenv>=1.0.0
requests>=2.31.0
aiohttp>=3.8.0
sentence-transformers>=2.2.0
numpy>=1.24.0
openai>=1.0.0
supabase>=2.0.0
httpx==0.25.0
backoff>=2.2.0
fastapi>=0.100.0
uvicorn>=0.23.0
